Renamed core-tracing-opentelemetry to core/tracing-opentelemetry
