+++ sdk/core/azure-core-tracing-opentelemetry/inc/azure/core/tracing-opentelemetry/rtti.hpp
