+++ sdk/core/azure-core-tracing-opentelemetry/src/opentelemetry.cpp
@@ -1,5 +1,5 @@
 
-#include "azure/core-tracing-opentelemetry/opentelemetry.hpp"
+#include "azure/core/tracing-opentelemetry/opentelemetry.hpp"
 #include <azure/core/nullable.hpp>
 #include <azure/core/tracing/tracing.hpp>
 #include <memory>
