+++ sdk/core/azure-core-tracing-opentelemetry/test/ut/service_support_test.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #define USE_MEMORY_EXPORTER 1
-#include "azure/core-tracing-opentelemetry/opentelemetry.hpp"
+#include "azure/core/tracing-opentelemetry/opentelemetry.hpp"
 #include "azure/core/internal/tracing/service_tracing.hpp"
 #include <azure/core/internal/json/json.hpp>
 #include <azure/core/test/test_base.hpp>
